+++ docs/examples/entangled_classes/example_02_overriding_methods.py
@@ -1,6 +1,6 @@
 # type: ignore
 from pygim.gimmicks import EntangledClass
-from pygim.exceptions import EntangledMethodError
+from pygim.explib import EntangledMethodError
 
 # This creates inheritable class locally so that it is more evident, which class is
 # used.
