+++ pygim-common/pyproject.toml
@@ -1,6 +1,6 @@
 [build-system]
-requires = ["setuptools ~= 58.0", "wheel", "pdm-backend"]
-build-backend = "pdm.backend"
+requires = ["setuptools ~= 58.0", "wheel", "pdm-backend", "pybind11", "toml"]
+#build-backend = "pdm.backend"
 
 [project]
 name = "pygim-common"
