+++ pygim-common/tests/examples/test_examples.py
@@ -17,6 +17,6 @@ def test_examples(module_path):
 
 
 if __name__ == "__main__":
-    from pygim.utils.testing import run_tests
+    from pygim.testing import run_tests
     from pygim.gimmicks import EntangledClass
     run_tests(__file__, module_name=EntangledClass.__module__, coverage=False)
